Check permission in Admin component
