+++ src/containers/NotFoundPage/NotFoundPage.module.scss
@@ -3,5 +3,5 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  height: 100%;
+  height: 100vh;
 }
